+++ app/src/main/java/com/boxdotsize/boxdotsize_android/retrofit/Response.kt
@@ -1,7 +1,7 @@
 package com.boxdotsize.boxdotsize_android.retrofit
 
 data class Response(
-    val height: Int,
-    val tall: Int,
-    val width: Int
+    val height: Float,
+    val tall: Float,
+    val width: Float
 )
